fixes to bit, bit64 and ff for R-3.0.3
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/ff@144 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/bit/src/bit.c
@@ -26,11 +26,11 @@ void bit_init(int   bits){
   LASTBIT = bits - 1;
   mask0 = calloc(BITS, sizeof(int));
   mask1 = calloc(BITS, sizeof(int));
-  int b = 1;
+  unsigned int b = 1;
   int i;
   for (i=0; i<BITS; i++){
-    mask1[i] = b;
-    mask0[i] = ~b;
+    mask1[i] = (int) b;
+    mask0[i] = (int) ~b;
     //Rprintf("i=%d mask0[i]=%d mask1[i]=%d\n", i, mask0[i], mask1[i]);
     b = b << 1;
   }
